update delete copyrights

+++ copyrights/src/controllers/projects/remove-project.controller.ts
@@ -1,4 +1,11 @@
-import { SuccessResponse, NotAllowedError, CopyRights, Project, MODELS } from '@duvdu-v1/duvdu';
+import {
+  SuccessResponse,
+  NotAllowedError,
+  CopyRights,
+  Project,
+  MODELS,
+  PERMISSIONS,
+} from '@duvdu-v1/duvdu';
 import { RequestHandler } from 'express';
 
 export const removeProjectHandler: RequestHandler<{ projectId: string }, SuccessResponse> = async (
@@ -15,6 +22,12 @@ export const removeProjectHandler: RequestHandler<{ projectId: string }, Success
     { isDeleted: true },
   );
   if (!project) return next(new NotAllowedError(undefined, req.lang));
+
+  if (
+    project.user.toString() !== req.loggedUser.id ||
+    req.loggedUser.role.permissions.includes(PERMISSIONS.removeCopyrightHandler)
+  )
+    return next(new NotAllowedError(undefined, req.lang));
 
   await Project.findOneAndDelete({
     project: {
